# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DistroMash meshes your Docker Distribution",
        "title": "DistroMash API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ahmet Kerem Aksoy",
            "email": "a.aksoy@tu-berlin.de"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/": {
            "get": {
                "description": "Get the home page.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Home"
                ],
                "summary": "get the home page",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/strategy/{name}": {
            "get": {
                "description": "Get the strategy.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Strategy"
                ],
                "summary": "get the strategy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of Strategy",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Strategy"
                        }
                    }
                }
            }
        },
        "/crdt": {
            "get": {
                "description": "Get all CRDT key-value pairs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crdt"
                ],
                "summary": "get all crdt key-value pairs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Crdt"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Post a CRDT key-value pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crdt"
                ],
                "summary": "post crdt key-value pair",
                "parameters": [
                    {
                        "description": "Post Crdt",
                        "name": "crdt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Crdt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Crdt"
                        }
                    }
                }
            }
        },
        "/crdt/{key}": {
            "get": {
                "description": "Get the CRDT value by key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crdt"
                ],
                "summary": "get crdt value by given key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key of Value",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Crdt"
                        }
                    }
                }
            }
        },
        "/strategy": {
            "get": {
                "description": "Get all strategies.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Strategy"
                ],
                "summary": "get all strategies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Strategy"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Post the strategy.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Strategy"
                ],
                "summary": "post the strategy",
                "parameters": [
                    {
                        "description": "Name of Strategy",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Strategy"
                        }
                    }
                }
            }
        },
        "/strategy/registered": {
            "get": {
                "description": "Get registered strategies.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Strategy"
                ],
                "summary": "get registered strategies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Strategy"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Crdt": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Strategy": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    }
}